PN-10326: runtime variables file creation

+++ cd-cli/commons/runtime-env-file-creation.sh
@@ -68,15 +68,22 @@ dump_params(){
   echo "Microservice Name:           ${microcvs_name}"
 }
 
+parse_params "$@"
+dump_params
+
 account_id=$(aws sts get-caller-identity --query Account --output text)
 bucket_env_path=${project_name}-runtime-environment-variables-${aws_region}-${account_id}
 file_env_name="runtime-variable.env"
 
-if aws s3api head-object --bucket ${bucket_env_path} --key ${microcvs_name}/${file_env_name} > /dev/null 2>&1; then
-    echo "File ${file_env_name} already exists."
+if aws s3api head-bucket --bucket ${bucket_env_path} 2>/dev/null; then 
+  if aws s3api head-object --bucket ${bucket_env_path} --key ${microcvs_name}/${file_env_name} > /dev/null 2>&1; then
+      echo "File ${file_env_name} already exists."
+  else
+    touch ./${file_env_name}
+    echo "Generating ${file_env_name}"
+    aws s3 cp ${file_env_name} s3://${bucket_env_path}/${microcvs_name}/
+    rm ./${file_env_name}
+  fi
 else
-  touch ./${file_env_name}
-  echo "Generating ${file_env_name}"
-  aws s3 cp ${file_env_name} s3://${bucket_env_path}/${microcvs_name}/
-  rm ./${file_env_name}
+  echo "Bucket ${bucket_env_path} does not exists."
 fi
